Stop forcing the top navbar to be always visible
Set the navbar-static-top class on the top navbar to make it scroll like
the rest of the webpage. Otherwise it's always shown at the top border
of the frame.

Given that the navbar on my page doesn't do much, it's unreasonable to
constantly use that much screen real estate.

See http://getbootstrap.com/components/#navbar for more documentation.